Remove the redundant PAGNode in PAGTree
